GH-237: Add styling arguments, improve plots, allow arbitrary variables
