+++ packages/plot/src/plot3d.rs
@@ -1,0 +1,12 @@
+/*
+
+use serde_json::{json, Value};
+use std::io::Write;
+use std::fs::File;
+use plotters::prelude::*;
+use plotters::style::ShapeStyle;
+use base64::{engine::general_purpose, Engine as _};
+
+use plot::{raw, get_calc_funcs, verify_function, eval_function};
+
+ */
